Converted YouTrackConfigurationForm to YouTrackConfiguration class that hands dictionaries that represent js fields to react through get_config
