+++ sentry_youtrack/forms.py
@@ -125,180 +125,3 @@ class DefaultFieldForm(forms.Form):
         default_fields[md5(force_bytes(data['field'], errors='replace')).hexdigest()] = data['value']
         self.plugin.set_option(
             self.plugin.default_fields_key, default_fields, self.project)
-
-
-class YouTrackConfigurationForm(forms.Form):
-
-    error_message = {
-        'client': _("Unable to connect to YouTrack."),
-        'project_unknown': _('Unable to fetch project'),
-        'project_not_found': _('Project not found: %s'),
-        'invalid_ssl': _("SSL certificate  verification failed."),
-        'invalid_password': _('Invalid username or password.'),
-        'invalid_project': _('Invalid project: \'%s\''),
-        'missing_fields': _('Missing required fields.'),
-        'perms': _("User doesn't have Low-level Administration permissions."),
-        'required': _("This field is required.")}
-
-    url = forms.URLField(
-        label=_("YouTrack Instance URL"),
-        widget=forms.TextInput(
-            attrs={'class': 'span9',
-                   'placeholder': 'e.g. "https://yoursitename.myjetbrains.com/youtrack/"'}),
-        required=True)
-    username = forms.CharField(
-        label=_("Username"),
-        help_text=_("User should have admin rights."),
-        widget=forms.TextInput(attrs={'class': 'span9'}),
-        required=True)
-    password = forms.CharField(
-        label=_("Password"),
-        help_text=_("Only enter a password if you want to change it"),
-        widget=forms.PasswordInput(attrs={'class': 'span9'}),
-        required=False)
-    project = forms.ChoiceField(
-        label=_("Linked Project"),
-        required=True)
-    default_tags = forms.CharField(
-        label=_("Default tags"),
-        help_text=_("Comma-separated list of tags"),
-        widget=forms.TextInput(
-            attrs={'class': 'span6', 'placeholder': "e.g. sentry"}),
-        required=False)
-    ignore_fields = forms.MultipleChoiceField(
-        label=_("Ignore fields"),
-        required=False,
-        help_text=_("These fields will not appear on the form"))
-
-    def __init__(self, *args, **kwargs):
-        super(YouTrackConfigurationForm, self).__init__(*args, **kwargs)
-        self.client_errors = {}
-
-        initial = kwargs.get("initial")
-        if initial:
-            client = self.get_youtrack_client(initial)
-            if not client:
-                self.remove_fields()
-            else:
-                if initial.get('project'):
-                    choices = self.get_ignore_field_choices(
-                        client, initial.get('project'))
-                    self.fields['ignore_fields'].choices = choices
-
-                choices = self.get_project_field_choices(
-                    client, initial.get('project'))
-                self.fields["project"].choices = choices
-
-                if not any(args) and not initial.get('project'):
-                    self.second_step_msg = _(
-                        "Your credentials are valid but plugin is NOT active "
-                        "yet. Please fill in remaining required fields.")
-
-            if self.client_errors and not args[0]:
-                self.full_clean()
-                for field, error in self.client_errors.items():
-                            self._errors[field] = [error]
-        else:
-            self.remove_fields()
-
-    def remove_fields(self):
-        del self.fields["project"]
-        del self.fields["default_tags"]
-        del self.fields["ignore_fields"]
-
-    def get_ignore_field_choices(self, client, project):
-        try:
-            fields = list(client.get_project_fields_list(project))
-        except HTTPError:
-            self.client_errors['project'] = self.error_message[
-                'invalid_project'] % (project,)
-        else:
-            names = [field['name'] for field in fields]
-            return zip(names, names)
-        return []
-
-    def get_project_field_choices(self, client, project):
-        choices = [(' ', u"- Choose project -")]
-        try:
-            projects = list(client.get_projects())
-        except HTTPError:
-            self.client_errors['project'] = self.error_message[
-                'invalid_project'] % (project, )
-        else:
-            for project in projects:
-                display = "%s (%s)" % (project['name'], project['id'])
-                choices.append((project['id'], display))
-        return choices
-
-    def get_youtrack_client(self, data, additional_params=None):
-        yt_settings = {
-            'url': data.get('url'),
-            'username': data.get('username'),
-            'password': data.get('password'),
-            'verify_ssl_certificate': VERIFY_SSL_CERTIFICATE}
-        if additional_params:
-            yt_settings.update(additional_params)
-
-        client = None
-        try:
-            client = YouTrackClient(**yt_settings)
-        except (HTTPError, ConnectionError) as e:
-            if e.response is not None and e.response.status_code == 403:
-                self.client_errors['username'] = self.error_message[
-                    'invalid_password']
-            else:
-                self.client_errors['url'] = self.error_message['client']
-        except (SSLError, TypeError) as e:
-            self.client_errors['url'] = self.error_message['invalid_ssl']
-        if client:
-            try:
-                client.get_user(yt_settings.get('username'))
-            except HTTPError as e:
-                if e.response.status_code == 403:
-                    self.client_errors['username'] = self.error_message['perms']
-                    client = None
-        return client
-
-    def get_project_fields_list(self, client, project_id):
-        try:
-            return list(client.get_project_fields_list(project_id))
-        except (HTTPError, ConnectionError) as e:
-            if e.response is not None and e.response.status_code == 404:
-                self.client_errors['project'] = self.error_message['project_not_found'] % project_id
-            else:
-                self.client_errors['project'] = self.error_message['project_unknown']
-
-    def get_projects(self, client):
-        try:
-            return list(client.get_projects())
-        except (HTTPError, ConnectionError) as e:
-            if e.response is not None and e.response.status_code == 404:
-                self.client_errors['project'] = self.error_message['project_not_found'] % project_id
-            else:
-                self.client_errors['project'] = self.error_message['project_unknown']
-
-    def clean_password(self):
-        password = (
-            self.cleaned_data.get('password') or self.initial.get('password'))
-        if not password:
-            raise ValidationError(self.error_message['required'])
-        return password
-
-    def clean_project(self):
-        project = self.cleaned_data.get('project').strip()
-        if not project:
-            raise ValidationError(self.error_message['required'])
-        return project
-
-    def clean(self):
-        data = self.cleaned_data
-        if not all(data.get(field)
-                   for field in ('url', 'username', 'password')):
-            raise ValidationError(self.error_message['missing_fields'])
-        client = self.get_youtrack_client(data)
-        if not client:
-            for field, error in self.client_errors.items():
-                self._errors[field] = [error]
-                if field in data.keys():
-                    del data[field]
-        return data
